create movie screen

+++ lib/models/movie_model.dart
@@ -19,7 +19,7 @@ class MovieModel {
       backdropPath: "https://image.tmdb.org/t/p/w500${map["backdrop_path"]}",
       posterPath: "https://image.tmdb.org/t/p/w500${map["poster_path"]}",
       title: map["title"],
-      voteAverage: map["vote_average"],
+      voteAverage: map["vote_average"].toDouble(),
     );
   }
 }
